+++ draft-packages/button/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@3.3.3...@kaizen/draft-button@3.3.4) (2021-05-11)
+
+**Note:** Version bump only for package @kaizen/draft-button
+
+
+
+
+
+## [3.3.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@3.3.2...@kaizen/draft-button@3.3.3) (2021-05-05)
+
+
+### Bug Fixes
+
+* use inline-flex for Button to fix edge case ([#1544](https://github.com/cultureamp/kaizen-design-system/issues/1544)) ([85758b1](https://github.com/cultureamp/kaizen-design-system/commit/85758b1584c5bef2c88dfb7a6a98cc89b36777b7))
+
+
+
+
+
+## [3.3.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@3.3.1...@kaizen/draft-button@3.3.2) (2021-05-05)
+
+**Note:** Version bump only for package @kaizen/draft-button
+
+
+
+
+
+## [3.3.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@3.3.0...@kaizen/draft-button@3.3.1) (2021-05-05)
+
+
+### Bug Fixes
+
+* add default variant for badge in button component when button is in reversed ([#1543](https://github.com/cultureamp/kaizen-design-system/issues/1543)) ([f9ff304](https://github.com/cultureamp/kaizen-design-system/commit/f9ff304efdf2242087ca5a2bc9ad3d08a2616dfa))
+
+
+
+
+
 # [3.3.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@3.2.18...@kaizen/draft-button@3.3.0) (2021-05-04)