+++ legacy-packages/title-block/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@3.1.2...@kaizen/draft-title-block@3.1.3) (2021-05-05)
+
+**Note:** Version bump only for package @kaizen/draft-title-block
+
+
+
+
+
 ## [3.1.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block@3.1.1...@kaizen/draft-title-block@3.1.2) (2021-05-04)
 
 **Note:** Version bump only for package @kaizen/draft-title-block